+++ src/app/cross/ui/components/avatar/src/Avatar.jsx
@@ -3,60 +3,13 @@
 import '../styles/desktop.scss'
 import {useRouter} from 'next/navigation'
 import Button from '../../button/src/Button'
-import {useEffect, useState} from 'react'
 import SloganAvatar from '../../sloganAvatar'
 import UserAvatar from '../../userAvatar'
-import 
-  getDataLocalStore,
- from '../../../../../cross/core/uses-cases/getDataLocalStore'
-import updateDataLocalStore from '../../../../../cross/core/uses-cases/updateDataLocalStore'
-import request from '../../../../core/uses-cases/request'
+import useAccount from '../../../../../cross/core/hoocks/useAccount/src/useAccount'
 
 const Avatar = () => {
-  const [isLogin, setIsLogin] = useState(false)
-  const [account, setAccount] = useState()
   const router = useRouter()
-
-
-   useEffect(() => {
-    setIsLogin(getDataLocalStore('token') === null ? false : true)
-  }, [])
-
-  const setAccountCallback = (data) => {
-    debugger
-    setAccount(data)
-  }
-
-  const callbackRequet = (data) => {
-    debugger
-    const {user_id} = data;
-    request({
-      path: `users/${user_id}`,
-      method: 'GET',
-      addHeaders: {
-        Authorization: getDataLocalStore('token'),
-      }
-    }, setAccountCallback)
-  }
-
-  useEffect(() => {
-    if (isLogin && !account) {
-      const dataRequest = {
-        path: 'account',
-        method: 'GET',
-        addHeaders: {
-          Authorization: getDataLocalStore('token'),
-        },
-      }
-      const r = request(dataRequest, callbackRequet)
-
-      if (r?.error) {
-        updateDataLocalStore('token', null)
-      }
-
-    } 
-
-  }, [isLogin, account])
+  const {account, isLogin} = useAccount()
 
   const handleSignUp = () => {
     router.push('/sign-email')
