+++ src/app/cross/ui/components/sloganAvatar/src/SloganAvatar.jsx
@@ -4,7 +4,6 @@ import '../styles/desktop.scss'
 const SloganAvatar = account => {
   const FIRSTLATTER = 0
   const {firstname = '', lastname = ''} = account ?? {}
-  debugger
   const firstLaterName = firstname?.charAt(FIRSTLATTER).toUpperCase()
   const firstLaterLastName = lastname?.charAt(FIRSTLATTER).toUpperCase()
   const defaultSlogan = 'contain-slogan'
